requirements.txt
ultralytics
opencv-python
streamlit
pdfplumber
edge-tts
moviepy
Pillow
pillow-heif
scikit-learn
numpy
